+++ changelog.md
@@ -1,6 +1,14 @@
 # Changelog
 
 
+## v1.1
+
+### What's Changed
+
+_(most recent changes are listed on top):_
+- Added configuration for modifying the Content-Security-Policy
+
+
 ## v1.0
 
 ### What's Changed